Create deletePhones.py
This script is designed to delete phone extensions in the VICIdial system one by one. When run, it prompts the user to enter a specific phone extension number to be deleted. After submitting the extension number, the script makes an API request to VICIdial to delete the specified phone extension. Upon completion of each deletion, the script asks the user if they wish to delete another extension. If the user responds with 'Y', the script continues and prompts for the next extension number. If 'N' is chosen, the script terminates.